Prune dead registry API
