+++ packages/webui-router/tests/fixtures/router-app/src/loading-skeleton/loading-skeleton.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { WebUIElement } from '@microsoft/webui-framework';
+
+export class LoadingSkeleton extends WebUIElement {}
