+++ packages/webui-router/tests/fixtures/router-app/src/page-alpha/page-alpha.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { WebUIElement } from '@microsoft/webui-framework';
+
+export class PageAlpha extends WebUIElement {}
