+++ packages/webui-router/tests/fixtures/router-app/src/page-beta/page-beta.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { WebUIElement } from '@microsoft/webui-framework';
+
+export class PageBeta extends WebUIElement {}
