+++ packages/webui-router/tests/fixtures/router-app/src/page-compose/page-compose.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { WebUIElement, attr } from '@microsoft/webui-framework';
+
+export class PageCompose extends WebUIElement {
+  @attr action = '';
+  @attr to = '';
+  @attr subject = '';
+}
