+++ packages/webui-router/tests/fixtures/router-app/src/page-detail/page-detail.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { WebUIElement, observable } from '@microsoft/webui-framework';
+
+export class PageDetail extends WebUIElement {
+  @observable itemId = '';
+}
